+++ CHANGELOG.md
@@ -25,7 +25,7 @@ Translations: [简体中文](CHANGELOG.zh.md)
 * depend: Update to jetbrains-lifecycle 2.8.4
 * depend: Update to kotlinx-coroutines 1.10.2
 * depend: Update to ktor2 2.3.13
-* depend: Update to ktor3 3.1.3
+* depend: Update to ktor3 3.0.3
 * depend: Update to okio 3.11.0
 * depend: Don't replace `kotlin-stdlib-jdk7` and `kotlin-stdlib-jdk8` with `kotlin-stdlib`